+++ R/finalfit_internal_functions.R
@@ -412,11 +412,13 @@ p_tidy = function(x, digits, prefix="="){
 #' @param n Value
 #' @param percent Value
 #' @param digits Value
+#' @param digits_n Value. Used when using weighted frequency counts
 #' @param na_include When proportion missing, include in parentheses?
 #'
 #' @export
 #'
-format_n_percent = function(n, percent, digits, na_include = TRUE) {
+format_n_percent = function(n, percent, digits, digits_n = 0, na_include = TRUE) {
+	n = round_tidy(n, digits_n)
 	percent = round_tidy(percent, digits)
 	out = paste0(n, " (", percent, ")")
 	if(!na_include){
@@ -734,7 +736,7 @@ finalfit_formula <- ff_formula
 #'
 #' @return A list containing dependent, explanatory and random effects variables 
 #' @export
-#'
+#' 
 #' @examples
 #' ff_parse_formula(mort ~ age + sex + (1 | hospital))
 ff_parse_formula <- function(.formula){
@@ -823,7 +825,7 @@ globalVariables(c("L95", "U95", "fit_id", "Total", "dependent",
 									"w", "Freq", "g", "total_prop", "Prop", "index_total", "vname", "Combined",
 									"2.5 %", "97.5 %", "p.value", "estimate", "index", "n", "missing_n", "var_type",
 									"missing_percent", "var1", "var2", "keep", "label", "rowid", "term",
-									"confint_L", "confint_U", "explanatory", "p"))
+									"confint_L", "confint_U", "explanatory", "p", "where"))
 
 
 # Workaround ::: as summary.formula not (yet) exported from Hmisc